Modularize app into utils and plots modules
